Fix Removing People
Removing didn't work well, in fact this fix is a hack. Trying to delete a connection from m_Connections in the Server class would result in a crash. The reason for the crash is unknown, but i'd be guessing m_Connections was being used somewhere else in the threading process. Anyways, m_Directory holds the list of the real connections.